Add files via upload
Music Player Using HTML and CSS
		 The Music Player is a simple project developed using JavaScript, CSS, and HTML. This project is an interesting project for listening to our self-made songs collection app. The user can play different kinds of music and enjoy the sound of their music. The user can click the forward and backward play button to change the songs.
This project is simply in HTML, CSS, and JavaScript. Taking about the features of this system, the user can listen to their music and relax by enjoying the music sound as their choice like the music player. This project includes a lot of JavaScript for making the functioning of the project.